+++ lorien/Main.gd
@@ -113,7 +113,7 @@ func _process(delta):
 # -------------------------------------------------------------------------------------------------
 func _handle_shortcut_actions() -> void:
 	if !_is_dialog_open():
-		if Input.is_action_just_pressed("copy_strokes") || Input.is_action_just_pressed("paste_strokes"):
+		if Input.is_action_just_pressed("copy_strokes") || Input.is_action_just_pressed("paste_strokes") || Input.is_action_just_pressed("duplicate_strokes"):
 			return
 
 		if Input.is_action_just_pressed("shortcut_new_project"):